requirements.txt
crace.egg==info
matplotlib==3.9.2
numpy==2.1.3
pandas==2.2.3
plotly==5.23.0
scikit_posthocs==0.10.0
seaborn==0.13.2
setuptools==74.1.2
statannotations==0.6.0
wheel==0.45.0
